requirements.txt
pymupdf==1.26.3
streamlit==1.46.1
openai==1.93.3
sentence-transformers==5.0.0
faiss-cpu==1.11.0